Move header component into correct folder
